requirements.txt
jittor
Pillow
numpy 
matplotlib
pyyaml
tensorboardX
opencv-python
tqdm
terminaltables
requests
scipy